Move fetchUpdatedDJs to /action/djs.js
