+++ react/frontpage/actions/djs.js
@@ -13,3 +13,24 @@ export const startFetching = () => ({
 export const stopFetching = () => ({
 	type: 'STOPPED_FETCHING'
 });
+
+/**
+Helpers
+**/
+const djsURL = '/api/djs';
+export const fetchUpdatedDJs = (dispatch) => {
+  dispatch(startFetching());
+  $.ajax({
+    url: djsURL,
+    dataType: 'json',
+    cache: false,
+    success: function(data) {
+      dispatch(stopFetching());
+      dispatch(updateDJs(data.djs));
+    }.bind(this),
+    error: function(xhr, status, err) {
+      dispatch(stopFetching());
+      console.error(djsURL, status, err.toString());
+    }.bind(this)
+  });
+}
