+++ react/frontpage/containers/DJsTab.jsx
@@ -3,10 +3,8 @@
 
 import { connect } from 'react-redux';
 
-import { updateDJs, startFetching, stopFetching } from '../actions/djs';
+import { updateDJs, startFetching, stopFetching, fetchUpdatedDJs } from '../actions/djs';
 import DJList from '../components/DJList.jsx';
-
-const djsURL = '/api/djs';
 
 const mapStateToProps = (state) => ({
 	djs: state.djs.djs,
@@ -24,25 +22,5 @@ const DJsTab = connect(
 	mapDispatchToProps
 )(DJList);
 
-/**
-Helpers
-**/
-
-const fetchUpdatedDJs = (dispatch) => {
-	dispatch(startFetching());
-  $.ajax({
-    url: djsURL,
-    dataType: 'json',
-    cache: false,
-    success: function(data) {
-    	dispatch(stopFetching());
-      dispatch(updateDJs(data.djs));
-    }.bind(this),
-    error: function(xhr, status, err) {
-    	dispatch(stopFetching());
-      console.error(djsURL, status, err.toString());
-    }.bind(this)
-  });
-}
 
 export default DJsTab;
